for #686: move ColumnDefinition41Packet to command package
