+++ sharding-proxy/src/main/java/io/shardingjdbc/proxy/backend/common/SQLExecuteBackendHandler.java
@@ -34,7 +34,7 @@ import io.shardingjdbc.proxy.constant.StatusFlag;
 import io.shardingjdbc.proxy.transport.common.packet.DatabaseProtocolPacket;
 import io.shardingjdbc.proxy.transport.mysql.packet.command.query.FieldCountPacket;
 import io.shardingjdbc.proxy.transport.mysql.packet.command.query.TextResultSetRowPacket;
-import io.shardingjdbc.proxy.transport.mysql.packet.generic.ColumnDefinition41Packet;
+import io.shardingjdbc.proxy.transport.mysql.packet.command.query.ColumnDefinition41Packet;
 import io.shardingjdbc.proxy.transport.mysql.packet.generic.EofPacket;
 import io.shardingjdbc.proxy.transport.mysql.packet.generic.ErrPacket;
 import io.shardingjdbc.proxy.transport.mysql.packet.generic.OKPacket;
